test: refactor EditorCommandHandlers-integ-test
